requirements.txt
Django==2.1.5
django-cors-headers==2.5.2
djangorestframework==3.9.1
pytz==2018.9
